LLM-Bot/requirements.txt
# ============================================================================
# LLM-Bot Requirements (compatible with env 'rag')
# Python 3.10+ required
# ============================================================================

# IMPORTANT: Install PyTorch with CUDA support FIRST via conda:
# conda install pytorch torchvision torchaudio pytorch-cuda=12.1 -c pytorch -c nvidia -y

# Core ML/DL
transformers>=4.36.0
accelerate>=0.25.0
sentencepiece>=0.1.99           # Required for T5 tokenization
protobuf>=3.20.0               # Required for T5

# Quantization (8-bit for flan-t5-xl)
bitsandbytes>=0.41.0

# PEFT/LoRA (for medical adapters - optional but recommended)
peft>=0.7.0

# RAG & Embeddings
langchain>=0.1.0
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
datasets>=2.16.0

# UI
gradio>=4.0.0
gradio-client>=0.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
tenacity>=8.2.0

# Scientific
numpy<2.0.0,>=1.21.0
pandas>=2.0.0

# Monitoring
tqdm>=4.65.0
psutil>=5.9.0

# System
colorama>=0.4.6
